removes duplicate declaration of GROUP_SIZE and QUORUM_SIZE

+++ src/types.rs
@@ -60,8 +60,8 @@ pub fn generate_random_vec_u8(size: usize) -> Vec<u8> {
     vec
 }
 
-pub static GROUP_SIZE: u32 = 32;
-pub static QUORUM_SIZE: u32 = 19;
+pub static GROUP_SIZE: usize = 32;
+pub static QUORUM_SIZE: usize = 19;
 
 pub trait Mergeable {
     fn merge<'a, I>(xs: I) -> Option<Self> where I: Iterator<Item=&'a Self>;
